[Florence2] Fix dtype mismatch in test_large_model_inference_eager
`florence-community/Florence-2-large` weights are stored in float16 on
the Hub. After PR #34919 made `torch_dtype="auto"` the default, the
model loads in float16 but processor outputs float32 tensors, causing:

    RuntimeError: Input type (float) and bias type (c10::Half) should be the same

Fix: pass `torch_dtype=torch.float32` explicitly to keep both the model
weights and inputs in float32, matching the original test intent.

Verified passing on A10 runner (torch 2.13.0+cu130).

+++ tests/models/florence2/test_modeling_florence2.py
@@ -448,9 +448,9 @@ class Florence2ForConditionalGenerationIntegrationTest(unittest.TestCase):
     def test_large_model_inference_eager(self):
         model_name = "florence-community/Florence-2-large"
         processor = AutoProcessor.from_pretrained(model_name)
-        model = Florence2ForConditionalGeneration.from_pretrained(model_name, attn_implementation="eager").to(
-            torch_device
-        )
+        model = Florence2ForConditionalGeneration.from_pretrained(
+            model_name, attn_implementation="eager", torch_dtype=torch.float32
+        ).to(torch_device)
 
         prompt = "<DETAILED_CAPTION>"
         inputs = processor(images=self.image1, text=prompt, return_tensors="pt")
